adaptive_engine/requirements.txt
pandas
numpy
websockets
supabase
python-dotenv
